crs/src/crs-cp-linux/fuzzer/SkyQEMU/requirements.txt
sphinx
pwntools
pyyaml
